+++ test/workspace/git_test.rb
@@ -34,19 +34,19 @@ class WorkspaceGitTest < Minitest::Test
 
   def test_remote_url
     with_workspace do |workspace|
-      assert_equal URI("https://github.com/sider/runners_test"), workspace.send(:remote_url)
+      assert_equal "https://github.com/sider/runners_test", workspace.send(:remote_url)
     end
   end
 
   def test_remote_url_with_file
     with_workspace(git_url: 'file:///project/smoke/foo/bar') do |workspace|
-      assert_equal URI("file:///project/smoke/foo/bar"), workspace.send(:remote_url)
+      assert_equal "file:///project/smoke/foo/bar", workspace.send(:remote_url)
     end
   end
 
   def test_remote_url_with_token
     with_workspace(git_url_userinfo: "x-access-token:v1.aaabbbcccdddeeefffgggbc06400127f248a82ac") do |workspace|
-      assert_equal URI("https://x-access-token:v1.aaabbbcccdddeeefffgggbc06400127f248a82ac@github.com/sider/runners_test"), workspace.send(:remote_url)
+      assert_equal "https://x-access-token:v1.aaabbbcccdddeeefffgggbc06400127f248a82ac@github.com/sider/runners_test", workspace.send(:remote_url)
     end
   end