+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 [Full diff](https://github.com/sider/runners/compare/0.39.0...HEAD)
 
 - Set `error` to Bugsnag severity [#1758](https://github.com/sider/runners/pull/1758)
+- Add `Runners::Workspace::Git::Error` class [#1759](https://github.com/sider/runners/pull/1759)
 
 ## 0.39.0